requirements.txt
Flask==3.1.1
flask-cors==6.0.0
gunicorn==21.2.0
google-generativeai==0.3.1
requests==2.32.3
colorama==0.4.6
python-dotenv==1.0.1